+++ src/commands/meson.build
@@ -1,0 +1,3 @@
+starbot_src += files(
+  './ping.c'
+)
