+++ src/meson.build
@@ -1,0 +1,5 @@
+subdir('./commands')
+
+starbot_src += files(
+  './main.c'
+)
